requirements.txt
pyserial>=3.5
supabase
python-dotenv
